requirements.txt
django
djangorestframework
llama-index
boto3
python-dotenv
openai
pdfplumber
python-docx
python-pptx
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
pinecone-client
openpyxl
tiktoken
llama-index-vector-stores-pinecone
ipython
llama_index
python-decouple
chardet
